pyproject.toml
[tool.poetry]
name = "cs4575-project2"
version = "0.1.0"
description = ""
authors = ["Reinier <reinier7175@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
seaborn = "^0.13.2"
numpy = "^2.2.3"
matplotlib = "^3.10.1"
pandas = "^2.2.3"
colorama = "^0.4.6"
docker = "^7.1.0"
pytz = "^2025.1"
transformers = "^4.49.0"
huggingface-hub = "^0.29.3"
llama-cpp-python = "^0.3.8"
deepspeed = "^0.16.4"
accelerate = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
